# Fix `clickhouse_query_perf.py` crash when all samples fail